python/pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["Rahul Chaki <rahulchaki@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
polars = "^1.12.0"
duckdb = "^1.1.2"
ibis-framework = {extras = ["duckdb", "polars"], version = "^9.5.0"}
faker = "^30.8.2"
mysql-connector-python = "^9.1.0"
connectorx = "^0.3.3"
sqlalchemy = "^2.0.36"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
